Added auth to discontinued controller and testing

+++ src/main/java/com/sparta/musicstoreapi/controllers/DiscontinuedController.java
@@ -2,8 +2,10 @@ package com.sparta.musicstoreapi.controllers;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sparta.musicstoreapi.entities.Discontinued;
+import com.sparta.musicstoreapi.entities.Token;
 import com.sparta.musicstoreapi.entities.Track;
 import com.sparta.musicstoreapi.repositories.DiscontinuedRepository;
+import com.sparta.musicstoreapi.repositories.TokenRepository;
 import com.sparta.musicstoreapi.repositories.TrackRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -22,20 +24,28 @@ public class DiscontinuedController {
     @Autowired
     private TrackRepository trackRepository;
     @Autowired
+    private TokenRepository tokenRepository;
+    @Autowired
     ObjectMapper mapper;
 
-    @PostMapping(value = "/track/discontinue/{trackId}", produces = { MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE, })
-    public ResponseEntity<?> discontinueTrack(@PathVariable Integer trackId, @RequestBody Discontinued discontinued) {
-        Optional<Track> trackResult = trackRepository.findById(trackId);
-        if (trackResult.isEmpty()) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Track not found!");
-        } else {
-            Track result = trackRepository.getById(trackId);
-            discontinued.setTrackId(result);
-            discontinuedRepository.save(discontinued);
-            Discontinued discontinuedResult = discontinuedRepository.getById(discontinued.getId());
-            return ResponseEntity.ok(discontinuedResult);
+    @PostMapping(value = "/track/discontinue/{trackId}/{token}", produces = { MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE, })
+    public ResponseEntity<?> discontinueTrack(@PathVariable Integer trackId, @RequestBody Discontinued discontinued, @PathVariable String token ) {
+        Optional<Token> tokenResult = tokenRepository.findByToken(token);
+        if (tokenResult.isPresent()) {
+            if (tokenResult.get().getPermissionLevel() >= 2) {
+                Optional<Track> trackResult = trackRepository.findById(trackId);
+                if (trackResult.isEmpty()) {
+                    return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Track not found!");
+                } else {
+                    Track result = trackRepository.getById(trackId);
+                    discontinued.setTrackId(result);
+                    discontinuedRepository.save(discontinued);
+                    Discontinued discontinuedResult = discontinuedRepository.getById(discontinued.getId());
+                    return ResponseEntity.ok(discontinuedResult);
+                }
+            }
         }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @GetMapping(value = "/track/discontinued/{trackId}", produces = { MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE, })
@@ -61,31 +71,43 @@ public class DiscontinuedController {
         return discontinuedRepository.findAllByIsDiscontinuedFalse();
     }
 
-    @DeleteMapping(value = "/track/discontinued/delete/{trackId}", produces = { MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE, })
-    public String deleteRow(@PathVariable Integer trackId) {
-        Track result = trackRepository.getById(trackId);
-        Integer toBeDeleted = discontinuedRepository.findByTrackId(result).getId();
-        discontinuedRepository.deleteById(toBeDeleted);
-        return "Row with : " + trackId + " deleted.";
+    @DeleteMapping(value = "/track/discontinued/delete/{trackId}/{token}", produces = { MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE, })
+    public String deleteRow(@PathVariable Integer trackId, @PathVariable String token) {
+        Optional<Token> tokenResult = tokenRepository.findByToken(token);
+        if (tokenResult.isPresent()) {
+            if (tokenResult.get().getPermissionLevel() >= 1) {
+                Track result = trackRepository.getById(trackId);
+                Integer toBeDeleted = discontinuedRepository.findByTrackId(result).getId();
+                discontinuedRepository.deleteById(toBeDeleted);
+                return "Row with : " + trackId + " deleted.";
+            }
+        }
+        return null;
     }
 
-    @PutMapping(value = "/track/discontinue/update/{trackId}", produces = { MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE })
-    public ResponseEntity<?> updateDiscontinued(@PathVariable Integer trackId, @RequestBody Discontinued newState) {
-        Optional<Track> trackResult = trackRepository.findById(trackId);
-        Track result = trackRepository.getById(trackId);
-        Integer id = discontinuedRepository.findByTrackId(result).getId();
-        newState.setId(id);
-        newState.setTrackId(result);
+    @PutMapping(value = "/track/discontinue/update/{trackId}/{token}", produces = { MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE })
+    public ResponseEntity<?> updateDiscontinued(@PathVariable Integer trackId, @RequestBody Discontinued newState, @PathVariable String token) {
+        Optional<Token> tokenResult = tokenRepository.findByToken(token);
+        if (tokenResult.isPresent()) {
+            if (tokenResult.get().getPermissionLevel() >= 1) {
+                Optional<Track> trackResult = trackRepository.findById(trackId);
+                Track result = trackRepository.getById(trackId);
+                Integer id = discontinuedRepository.findByTrackId(result).getId();
+                newState.setId(id);
+                newState.setTrackId(result);
 
-        if (trackResult.isEmpty()) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Track not found!");
-        } else {
-            Optional<Discontinued> oldState = discontinuedRepository.findById(newState.getId());
-            if (oldState.isEmpty()) {
-                return null;
+                if (trackResult.isEmpty()) {
+                    return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Track not found!");
+                } else {
+                    Optional<Discontinued> oldState = discontinuedRepository.findById(newState.getId());
+                    if (oldState.isEmpty()) {
+                        return null;
+                    }
+                    discontinuedRepository.save(newState);
+                    return ResponseEntity.ok(newState);
+                }
             }
-            discontinuedRepository.save(newState);
-            return ResponseEntity.ok(newState);
         }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 }
